Componentize some emails
