+++ packages/react-email-templates/emails/session-insights.tsx
@@ -1,18 +1,15 @@
 import {
-	Body,
 	Column,
-	Container,
-	Head,
 	Heading,
 	Hr,
-	Html,
 	Img,
 	Link,
-	Preview,
 	Section,
 	Text,
 } from '@react-email/components'
 import * as React from 'react'
+
+import { EmailHtml, HighlightLogo } from '../components/common'
 
 interface InterestingSession {
 	identifier: string
@@ -51,13 +48,6 @@ const sessionExample = {
 	],
 }
 
-const css = `
-    a {
-        color: unset;
-		text-decoration: none;
-    }
-`
-
 export const SessionInsightsEmail = ({
 	projectName = 'Highlight Production (app.highlight.io)',
 	toEmail = 'zane@highlight.io',
@@ -65,176 +55,138 @@ export const SessionInsightsEmail = ({
 	useHarold = false,
 	interestingSessions = [sessionExample, sessionExample, sessionExample],
 }: SessionInsightsEmailProps) => (
-	<Html>
-		<Head>
-			<style>{css}</style>
-		</Head>
-		<Preview>Session insights for {projectName}</Preview>
-		<Body style={main}>
-			<Container width={400} style={container}>
-				<Img
-					src="https://static.highlight.io/assets/digest/logo-on-dark.png"
-					width="32"
-					height="32"
-					alt="Highlight logo"
-					style={logo}
-				/>
-				<Heading style={headingText}>
-					<span style={highlightedText}>
-						<a style={highlightedText}>{projectName}</a>
-					</span>
-					<br />
-					Session Insights
-				</Heading>
-				<Text style={subtitleText}>
-					Here are 3 interesting* sessions recorded in your project
-					this week:
-				</Text>
-				<Hr style={hr} />
-				{interestingSessions.map((s, idx) => (
-					<>
-						<Section width={378}>
-							<Link href={s.url}>
-								<Img
-									src={s.screenshotUrl}
-									style={sessionScreenshot}
-									width={378}
-									height={206}
-								/>
+	<EmailHtml previewText={`Session insights for ${projectName}`}>
+		<HighlightLogo />
+		<Heading style={headingText}>
+			<span style={highlightedText}>
+				<a style={highlightedText}>{projectName}</a>
+			</span>
+			<br />
+			Session Insights
+		</Heading>
+		<Text style={subtitleText}>
+			Here are 3 interesting* sessions recorded in your project this week:
+		</Text>
+		<Hr style={hr} />
+		{interestingSessions.map((s, idx) => (
+			<>
+				<Section width={378}>
+					<Link href={s.url}>
+						<Img
+							src={s.screenshotUrl}
+							style={sessionScreenshot}
+							width={378}
+							height={206}
+						/>
+					</Link>
+				</Section>
+				<Section width={378} style={sessionAttributes}>
+					<Column align="left" width={266} style={sessionProperties}>
+						<Section align="left" style={leftAlign}>
+							<Text style={identifier}>
+								{s.avatarUrl && (
+									<>
+										<Img
+											src={s.avatarUrl}
+											width="22"
+											height="22"
+											style={avatar}
+										/>
+										&nbsp;
+									</>
+								)}
+								<a style={identifierText}>{s.identifier}</a>
+							</Text>
+						</Section>
+						<Section align="left" style={leftAlign}>
+							<span style={sessionProperty}>{s.country}</span>
+							&nbsp;
+							<span style={sessionProperty}>
+								{s.activeLength}
+							</span>
+						</Section>
+						<Section
+							align="left"
+							style={{ ...leftAlign, marginTop: '8px' }}
+						>
+							<Link style={viewSessionText} href={s.url}>
+								View Session →
 							</Link>
 						</Section>
-						<Section width={378} style={sessionAttributes}>
-							<Column
-								align="left"
-								width={266}
-								style={sessionProperties}
-							>
-								<Section align="left" style={leftAlign}>
-									<Text style={identifier}>
-										{s.avatarUrl && (
-											<>
-												<Img
-													src={s.avatarUrl}
-													width="22"
-													height="22"
-													style={avatar}
-												/>
-												&nbsp;
-											</>
-										)}
-										<a style={identifierText}>
-											{s.identifier}
-										</a>
-									</Text>
-								</Section>
-								<Section align="left" style={leftAlign}>
-									<span style={sessionProperty}>
-										{s.country}
-									</span>
-									&nbsp;
-									<span style={sessionProperty}>
-										{s.activeLength}
-									</span>
-								</Section>
-								<Section
-									align="left"
-									style={{ ...leftAlign, marginTop: '8px' }}
-								>
-									<Link style={viewSessionText} href={s.url}>
-										View Session →
-									</Link>
-								</Section>
-							</Column>
-							<Column width={132} style={activityGraphColumn}>
-								<Img
-									style={activityGraph}
-									src={s.activityGraphUrl}
-									width="132"
-									height="66"
-									alt="Session activity graph"
-								/>
-							</Column>
-						</Section>
-						{s.insights.map((i, idx) => (
-							<Section width={378} key={idx} style={insight}>
-								<Column width={24} style={numberLabel}>
-									{idx + 1}
-								</Column>
-								<Column
-									width={354}
-									style={{ wordBreak: 'break-all' }}
-								>
-									<Text style={insightText}>{i}</Text>
-								</Column>
-							</Section>
-						))}
-						<Hr style={hr} />
-					</>
+					</Column>
+					<Column width={132} style={activityGraphColumn}>
+						<Img
+							style={activityGraph}
+							src={s.activityGraphUrl}
+							width="132"
+							height="66"
+							alt="Session activity graph"
+						/>
+					</Column>
+				</Section>
+				{s.insights.map((i, idx) => (
+					<Section width={378} key={idx} style={insight}>
+						<Column width={24} style={numberLabel}>
+							{idx + 1}
+						</Column>
+						<Column width={354} style={{ wordBreak: 'break-all' }}>
+							<Text style={insightText}>{i}</Text>
+						</Column>
+					</Section>
 				))}
-				<Text style={paragraph}>
-					* These are sessions with unusual user journeys. You can
-					read more about our methodology{' '}
-					<Link
-						style={anchor}
-						href="https://www.highlight.io/blog/interesting-sessions"
-					>
-						here
-					</Link>
-					.
-				</Text>
-				{!useHarold && (
-					<Text style={paragraph}>
-						Your workspace has AI insights turned off. If you would
-						like this digest to include a summary of each session,
-						you can turn on AI insights{' '}
-						<Link
-							style={anchor}
-							href="https://app.highlight.io/w/harold-ai"
-						>
-							here
-						</Link>
-						.
-					</Text>
-				)}
-				<Text style={paragraph}>
-					This digest was sent to{' '}
-					<Link style={anchor} href={`mailto:${toEmail}`}>
-						{toEmail}
-					</Link>
-					. If you don't want to receive emails like this in the
-					future, you can{' '}
-					<Link style={anchor} href={unsubscribeUrl}>
-						unsubscribe
-					</Link>
-					.
-				</Text>
 				<Hr style={hr} />
-				<Img
-					style={logoFull}
-					src="https://static.highlight.io/assets/digest/highlight-logo.png"
-					width="70"
-					height="16"
-					alt="Highlight logo"
-				/>
-				<Text style={footer}>Seattle, WA 98122</Text>
-			</Container>
-		</Body>
-	</Html>
+			</>
+		))}
+		<Text style={paragraph}>
+			* These are sessions with unusual user journeys. You can read more
+			about our methodology{' '}
+			<Link
+				style={anchor}
+				href="https://www.highlight.io/blog/interesting-sessions"
+			>
+				here
+			</Link>
+			.
+		</Text>
+		{!useHarold && (
+			<Text style={paragraph}>
+				Your workspace has AI insights turned off. If you would like
+				this digest to include a summary of each session, you can turn
+				on AI insights{' '}
+				<Link
+					style={anchor}
+					href="https://app.highlight.io/w/harold-ai"
+				>
+					here
+				</Link>
+				.
+			</Text>
+		)}
+		<Text style={paragraph}>
+			This digest was sent to{' '}
+			<Link style={anchor} href={`mailto:${toEmail}`}>
+				{toEmail}
+			</Link>
+			. If you don't want to receive emails like this in the future, you
+			can{' '}
+			<Link style={anchor} href={unsubscribeUrl}>
+				unsubscribe
+			</Link>
+			.
+		</Text>
+		<Hr style={hr} />
+		<Img
+			style={logoFull}
+			src="https://static.highlight.io/assets/digest/highlight-logo.png"
+			width="70"
+			height="16"
+			alt="Highlight logo"
+		/>
+		<Text style={footer}>Seattle, WA 98122</Text>
+	</EmailHtml>
 )
 
 export default SessionInsightsEmail
-
-const main = {
-	backgroundColor: '#0d0225',
-	fontFamily: 'Helvetica, sans-serif',
-}
-
-const container = {
-	width: '400px',
-	padding: '0 16px',
-	textAlign: 'center' as const,
-	margin: '0 auto',
-}
 
 const hr = {
 	color: '#30294e',
@@ -258,11 +210,6 @@ const footer = {
 	color: '#9d97aa',
 	fontSize: '12px',
 	lineHeight: '16px',
-}
-
-const logo = {
-	paddingTop: '32px',
-	margin: '0 auto',
 }
 
 const logoFull = {
